+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClassController;
 use App\Http\Controllers\ExtracurricularController;
 use App\Http\Controllers\StudentController;
@@ -18,50 +19,52 @@ use Illuminate\Support\Facades\Route;
  */
 
 Route::get('/', function () {
-    return view('home', [
-        'name' => 'mamad',
-        'role' => 'admin',
-        'buah' => ['nanas', 'jambu', 'rujak'],
-    ]);
-});
+    return view('home');
+})->middleware('auth');
+
+//halaman login
+Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
+Route::post('/login', [AuthController::class, 'otentikasi'])->middleware('guest');
+Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth');
+
 //student
-Route::get('/students', [StudentController::class, 'index']);
-Route::get('/student/{id}', [StudentController::class, 'show']);
-Route::get('/addstudent', [StudentController::class, 'create']);
-Route::post('/student', [StudentController::class, 'store']);
-Route::get('/editstudent/{id}', [StudentController::class, 'edit']);
-Route::put('/student-edit/{id}', [StudentController::class, 'update']);
-Route::get('/deletestudent/{id}', [StudentController::class, 'delete']);
-Route::delete('/student-remove/{id}', [StudentController::class, 'remove']);
-Route::get('/student-restore', [StudentController::class, 'restoreStudent']);
-Route::get('/student/{id}/restore', [StudentController::class, 'restore']);
+Route::get('/students', [StudentController::class, 'index'])->middleware('auth');
+Route::get('/student/{id}', [StudentController::class, 'show'])->middleware(['auth', 'only-admin-teacher']);
+Route::get('/addstudent', [StudentController::class, 'create'])->middleware(['auth', 'only-admin-teacher']);
+Route::post('/student', [StudentController::class, 'store'])->middleware(['auth', 'only-admin-teacher']);
+Route::get('/editstudent/{id}', [StudentController::class, 'edit'])->middleware(['auth', 'only-admin-teacher']);
+Route::put('/student-edit/{id}', [StudentController::class, 'update'])->middleware(['auth', 'only-admin-teacher']);
+Route::get('/deletestudent/{id}', [StudentController::class, 'delete'])->middleware(['auth', 'only-admin']);
+Route::delete('/student-remove/{id}', [StudentController::class, 'remove'])->middleware(['auth', 'only-admin']);
+Route::get('/student-restore', [StudentController::class, 'restoreStudent'])->middleware(['auth', 'only-admin']);
+Route::get('/student/{id}/restore', [StudentController::class, 'restore'])->middleware(['auth', 'only-admin']);
 
 //class
-Route::get('/class', [ClassController::class, 'index']);
-Route::get('/class/{id}', [ClassController::class, 'show']);
-Route::get('/addclass', [ClassController::class, 'create']);
-Route::post('/class-add', [ClassController::class, 'store']);
-Route::get('/editclass/{id}', [ClassController::class, 'edit']);
-Route::put('/class-edit/{id}', [ClassController::class, 'update']);
+Route::get('/class', [ClassController::class, 'index'])->middleware('auth');
+Route::get('/class/{id}', [ClassController::class, 'show'])->middleware('auth');
+Route::get('/addclass', [ClassController::class, 'create'])->middleware('auth');
+Route::post('/class-add', [ClassController::class, 'store'])->middleware('auth');
+Route::get('/editclass/{id}', [ClassController::class, 'edit'])->middleware('auth');
+Route::put('/class-edit/{id}', [ClassController::class, 'update'])->middleware('auth');
 
 //ekskul
-Route::get('/extracurricular', [ExtracurricularController::class, 'index']);
-Route::get('/extracurricular/{id}', [ExtracurricularController::class, 'show']);
-Route::get('/addextracurricular', [ExtracurricularController::class, 'create']);
-Route::post('/extracurricular-add', [ExtracurricularController::class, 'store']);
-Route::get('/editextracurricular/{id}', [ExtracurricularController::class, 'edit']);
-Route::put('/extracurricular-add/{id}', [ExtracurricularController::class, 'update']);
+Route::get('/extracurricular', [ExtracurricularController::class, 'index'])->middleware('auth');
+Route::get('/extracurricular/{id}', [ExtracurricularController::class, 'show'])->middleware('auth');
+Route::get('/addextracurricular', [ExtracurricularController::class, 'create'])->middleware('auth');
+Route::post('/extracurricular-add', [ExtracurricularController::class, 'store'])->middleware('auth');
+Route::get('/editextracurricular/{id}', [ExtracurricularController::class, 'edit'])->middleware('auth');
+Route::put('/extracurricular-add/{id}', [ExtracurricularController::class, 'update'])->middleware('auth');
 
 //anggota ekskul
-Route::get('/addanggota', [ExtracurricularController::class, 'createanggota']);
-Route::post('/anggota-add', [ExtracurricularController::class, 'storeanggota']);
-Route::get('/editanggota/{id}', [ExtracurricularController::class, 'editanggota']);
-Route::put('/anggota-edit/{id}', [ExtracurricularController::class, 'updateanggota']);
+Route::get('/addanggota', [ExtracurricularController::class, 'createanggota'])->middleware('auth');
+Route::post('/anggota-add', [ExtracurricularController::class, 'storeanggota'])->middleware('auth');
+Route::get('/editanggota/{id}', [ExtracurricularController::class, 'editanggota'])->middleware('auth');
+Route::put('/anggota-edit/{id}', [ExtracurricularController::class, 'updateanggota'])->middleware('auth');
 
 //teacher
-Route::get('/teacher', [TeacherController::class, 'index']);
-Route::get('/teacher/{id}', [TeacherController::class, 'show']);
-Route::get('/addteacher', [TeacherController::class, 'create']);
-Route::post('/teacher-add', [TeacherController::class, 'store']);
-Route::get('/editteacher/{id}', [TeacherController::class, 'edit']);
-Route::put('/teacher-add/{id}', [TeacherController::class, 'update']);
+Route::get('/teacher', [TeacherController::class, 'index'])->middleware('auth')->middleware('auth');
+Route::get('/teacher/{id}', [TeacherController::class, 'show'])->middleware('auth');
+Route::get('/addteacher', [TeacherController::class, 'create'])->middleware('auth');
+Route::post('/teacher-add', [TeacherController::class, 'store'])->middleware('auth');
+Route::get('/editteacher/{id}', [TeacherController::class, 'edit'])->middleware('auth');
+Route::put('/teacher-add/{id}', [TeacherController::class, 'update'])->middleware('auth');
